requirements.txt
flask
nodejs
requests
python-dotenv
openai
gtts
mutagen